refactor(widgets): unify cursor-cycle helper across all list pickers
`agent_picker`, `github_picker`, `workdir_pick` each carried a
hand-rolled wrap-around cursor body inside their Up/Down arms — six
near-identical clones of the `(cur == 0 ? n-1 : cur-1)` /
`(cur+1 >= n ? 0 : cur+1)` pattern. `op_picker` had its own private
`Self::step_selection` helper covering the same four-pane case.

Hoist the helper to `console::widgets::cycle_select` (pub(crate)), drop
op_picker's private copy, and route all five pickers through it. The
`unwrap_or(0)` vs `map_or(0, …)` difference was non-load-bearing: every
caller already initializes `select(Some(0))` on a non-empty list and the
`count == 0` short-circuit covers the empty case identically.

Net: -45 LOC, single source of truth for wrap semantics, zero
behaviour change.
