+++ src/console/widgets/agent_picker.rs
@@ -55,25 +55,11 @@ impl AgentPickerState {
     }
 
     fn move_up(&mut self) {
-        let n = self.filtered.len();
-        if n > 0 {
-            let next = self
-                .list_state
-                .selected
-                .map_or(0, |i| if i == 0 { n - 1 } else { i - 1 });
-            self.list_state.select(Some(next));
-        }
+        super::cycle_select(&mut self.list_state, self.filtered.len(), -1);
     }
 
     fn move_down(&mut self) {
-        let n = self.filtered.len();
-        if n > 0 {
-            let next = self
-                .list_state
-                .selected
-                .map_or(0, |i| if i + 1 >= n { 0 } else { i + 1 });
-            self.list_state.select(Some(next));
-        }
+        super::cycle_select(&mut self.list_state, self.filtered.len(), 1);
     }
 
     pub fn handle_key(&mut self, key: KeyEvent) -> ModalOutcome<ClassSelector> {
